Made new folder for pygame model.added acii art.Fixed spelling in readm
