+++ art.py
@@ -1,0 +1,58 @@
+hangman = [
+'''
+   _ _
+  |   |
+      |
+      |
+      |
+      |
+==========
+''',
+'''
+   _ _
+  |   |
+  O   |
+      |
+      |
+      |
+========== ''',
+'''
+   _ _
+  |   |
+  O   |
+  |   |
+      |
+      |
+==========''',
+'''
+   _ _
+  |   |
+  O   |
+ /|   |
+      |
+      |
+==========''',
+'''
+   _ _
+  |   |
+  O   |
+ /|\  |
+      |
+      |
+==========''',
+'''
+   _ _
+  |   |
+  O   |
+ /|\  |
+ /    |
+      |
+==========''',
+'''
+   _ _
+  |   |
+  O   |
+ /|\  |
+ / \  |
+      |
+==========''']
